tests/e2e/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
playwright==1.40.0
python-dotenv==1.0.0
requests==2.31.0